BinderPlots.BinderPlots — Module

BinderPlots is a lightweight interface to the underlying PlotlyLight package (Fastest time-to-first-plot in Julia!), itself an "ultra-lightweight interface" to the Plotly javascript libraries. BinderPlots supports some of the Plots.jl interface for plotting. This package may be of use as an alternative to Plots in resource-constrained environments, such as binder.

Plots.jl uses

positional arguments for data

keyword arguments for attributes

plot() is a workhorse with seriestype indicating which plot; there are also special methods (e.g. scatter(x,y) becomes plot(x,y; seriestype="scatter"))

PlotlyLight.jl uses

data::Vector{Config} to hold tracts of data for plotting

layout::Config to adjust layout

config::Config to adjust global configurations

Config very flexibly creates the underlying Javascript objects the plotly interface expects

Plot() is a workhorse with type acting like seriestype and also mode

BinderPlots.jl has this dispatch for plot:

Line plot. connecting x,y (and possibly z). For 2D, use !isfinite values in y to break.

plot(x,y,[z]; kwargs...)
plot!([p::Plot], x, y, [z]; kwargs...)
plot(pts; kwargs...)
plot!([p::Plot], pts; kwargs...)

Data can be generated from a function:

plot(f::Function, ab; kwargs...) => plot(unzip(f, ab)...; kwargs...)
plot(f::Function, a, b; kwargs...) => plot(unzip(f, a, b)...; kwargs...)
plot!([p::Plot], f, ab, [b])

plot each function as lineplot:

plot(fs::Vector{Function}, a, [b]; kwargs...)
plot!([p::Plot], fs::Vector{Function}, a, [b]; kwargs...)

Note

Currently x, y make vectors; should matrices be supported using column vectors? "In Plots.jl, every column is a series, a set of related points which form lines, surfaces, or other plotting primitives. "

Parametric line plots, 2 or 3d

plot(fs::NTuple(N,Function), a, [b]; kwargs...)
plot!([p::Plot], fs::NTuple(N,Function), a, [b]; kwargs...)

Alternatively

plot(u::Function, v::Function, [w::Function], a, [b]; kwargs...)

The plot interface of PlotlyLight: merge layout; merge config; pass kwargs to Config push onto data or merge onto last tract:

plot(; layout::Config?, config::Config, kwargs...)
plot!([p::Plot]; layout::Config?, config::Config?, kwargs...)

This interface can be used to generate other plot types either by specifying the type argument, or using the form plot.plot_type(...), as with plot.scatter(x, y).

The plot function primarily plots line plots where the specified points are connected with lines (when finite); The scatter function plots just the points.

In addition there are these plot constructors for higher-dimensional plots

contour

plot_implicit

heatmap

surface

wireframe

There are also numerous functions to modify attributes of an existing plot.

BinderPlots.BezierCurve — Type

create a BezierCurve for plotting

From Plots.jl

BinderPlots.Series — Type

Series(xs₁, xs₂, …)

Struct to indicate traces in a series. Can be used in place of matrix to combine mismatched sizes, e.g.

scatter(BinderPlots.Series(1:3, 1:5), markersize=(20,10))

Not exported.

BinderPlots.Shape — Type

Shape(x, y)
Shape(vertices)

Construct a polygon to be plotted.

When plotting shapes, use stroke argument to fill to adjust bounding line properties.

BinderPlots.SeriesType — Method

SeriesType(x::Symbol)

Return a type and mode for plotly based on a Plots.jl series type.

BinderPlots.abline! — Method

abline!([p::Plot], intercept, slope; kwargs...)

Draw line y = a + bx over current viewing window, as determined by extrema(p).

BinderPlots.annotate! — Method

annotate!([p::Plot], x, y, txt; [color], [family], [pointsize], [halign], [valign])
annotate!([p::Plot], anns::Tuple;  kwargs...)

Add annotations to plot.

x, y, txt: text to add at (x,y)

color: text color

family: font family

pointsize: text size

halign: one of "top", "bottom"

valign: one of "left", "right"

rotation: angle to rotate

The x, y, txt values can be specified as 3 iterables or tuple of tuples.

BinderPlots.arrow — Method

arrow(tails, vs; kwargs...)

Draw vectors vs anchored at tails. Hacked in support for 3D using combination of lines + cones.

Use Plotly attributes arrowcolor, arrowwidth,

BinderPlots.arrows — Method

arrows(x, y, [z], f::Function; kwargs...)
arrows(x, y, [z], u, v, [w]; kwargs...)
arrows!([p::Plot], x, y, [z], f::Function; kwargs...)
arrows!([p::Plot], x, y, [z], u, v, [w]; kwargs...)

Draw arrows from x-y-[z] with lengths u-v-w (or optionally generated by f).

Example

plot(;xlims=(-5,5), ylims=(-5,5))
f(x,y) = (-y,x)./sqrt(1 + x^2 + y^2)
grid = tuple.((-5:5)', (-5:5))
arrows!(unzip(grid)..., f)

BinderPlots.band — Method

band(lower, upper; kwargs...)
band(lower::Function, upper::Function, a::Real, b::Real,n=251; kwargs...)
band!([p::Plot],lower, upper; kwargs...)
band!([p::Plot],lower::Function, upper::Function, a::Real, b::Real,n=251; kwargs...)

Draw band between lower and upper. These may be specified by functions or by tuples of x-y-[z] values.

Example

Using (x,y) points to define the boundaries

xs = 1:0.2:10
ys_low = -0.2 .* sin.(xs) .- 0.25
ys_high = 0.2 .* sin.(xs) .+ 0.25

p = plot(;xlims=(0,10), ylims=(-1.5, .5), legend=false)
band!(zip(xs, ys_low), zip(xs, ys_high); fillcolor=:blue)
band!(zip(xs, ys_low .- 1), zip(xs, ys_high .- 1); fillcolor=:red)

Or, using functions to define the boundaries

band(x -> -0.2 * sin(x) - 0.25, x -> 0.2 * sin(x) + 0.25,
     0, 10;  # a, b, n=251
     fillcolor=:red, legend=false)

BinderPlots.center — Method

return the centroid of a Shape

BinderPlots.circ3d — Method

circ3d(q, r, n̄; kwargs...)
circ3d!([p::Plot], q, r, n̄; kwargs...)

Plot circle in 3 dimensions with center at q, radius r, and perpendicular to normal vector n̄.

Example

q, n = [0,0,0], [0,0,1]
circ3d(q, 3, n)
arrow!(q, n)

Or a more complicated one:

Z(r, θ) = 4 - r
X(r, θ) = r * cos(θ)
Y(r, θ) = r * sin(θ)
rs = range(0,4, length=10)
θs = range(0, 2pi, length=100)
surface(X.(rs', θs), Y.(rs', θs), Z.(rs', θs); opacity=0.25)
q = [0,0,2]
n = [0,0,1]
r = 2
circ3d!(q, r, n; color="black", opacity=0.75)
q, v̂, ŵ = [0,0,0], [0,4,0], [0,0,4]
parallelogram!(q, v̂, ŵ; opacity=0.5, color=:yellow)

BinderPlots.circle! — Method

circle([p::Plot], x0, x1, y0, y1; kwargs...)

Draw circle shape bounded in [x0, x1] × [y0, y1]. (Will adjust to non-equal sized boundary.)

Example

Use named tuple for line for boundary.

circle!(p, 2,3,-1,1; line=(color=:gray,), fillcolor=:red, opacity=0.2)

BinderPlots.contour — Method

contour(x, y, z; kwargs...)
contour!([p::Plot], x, y, z; kwargs...)
contour(x, y, f::Function; kwargs...)
contour!(x, y, f::Function; kwargs...)
contourf(x, y, z; kwargs...)
contourf!([p::Plot], x, y, z; kwargs...)
contourf(x, y, f::Function; kwargs...)
contourf!(x, y, f::Function; kwargs...)

Create contour map.

Use contourf variants for filled contour plot.

Use seriescolor or palette to pass different colorschemes, e.g. seriescolor=cgrad(:acton).

Example

f(x,y) = sinpi(x*y)*sinpi(x/2 + y)
ys = xs = range(-1,1,200)
contourf(xs, ys, f; seriescolor=cgrad(:algae))

BinderPlots.current — Method

current()

Get current figure. A Plot object of PlotlyLight; UndefRefError if none.

Not typically needed, as it is implicit in most mutating calls, though may be convenient if those happen within a loop.

BinderPlots.default_size! — Method

default_size!(w::Integer,h::Integer)

Specify default size as percent of allotted screen.

PlotlyLight uses 100, 100 which looks odd in a Jupyter notebook. We set this to 60,120 (so width is 60% of screen size), which seems to look better.

BinderPlots.font — Method

font(args...)

(This is from Plots.jl)

Create a Font from a list of features. Values may be specified either as arguments (which are distinguished by type/value) or as keyword arguments.

Arguments

family: AbstractString. "serif" or "sans-serif" or "monospace"

pointsize: Integer. Size of font in points

halign: Symbol. Horizontal alignment (:hcenter, :left, or :right)

valign: Symbol. Vertical alignment (:vcenter, :top, or :bottom)

rotation: Real. Angle of rotation for text in degrees (use a non-integer type). (Works with ticks and annotations.)

color

Examples

julia> font(8)
julia> font(family="serif", halign=:center, rotation=45.0)

BinderPlots.grid_layout — Method

grid_layout(ps::Array{<:Plot})

Layout an array of plots into a grid. Vectors become rows of plots.

Use Plot() to create an empty plot for a given cell.

Example

using DataFrames
n = 25
d = DataFrame(x=sin.(rand(n)), y=rand(n).^2, z = rand(n)) # assume all numeric
nms = names(d)
m = Matrix{Plot}(undef, length(nms), length(nms))

for i ∈ eachindex(nms)
    for j ∈ eachindex(nms)
        if j > i
            p = Plot()
        elseif j == i
            x = d[:,j]
            p = plot(x, seriestype=:histogram)
            xlabel!(p, nms[j])  # <<- not working in grid! as differe xaxis purposes
        else
            x = d[:,i]; y = d[:,j]
            p = scatter(x, y)
            xlabel!(p, nms[i]); ylabel!(p, nms[j])
        end
        m[i,j] = p
    end
end
plot(m)

BinderPlots.heatmap — Method

heatmap(x, y, z; kwargs...)
heatmap!([p::Plot], x, y, z; kwargs...)
heatmap(x, y, f::Function; kwargs...)
heatmap!(x, y, f::Function; kwargs...)

Create heatmap function of f

BinderPlots.hline! — Method

hline!(y; xmin=0, xmax=1.0; kwargs...)

Draw horizontal line at y. By default extends over the current plot range, this can be adjusted by xmin and xmax, values in [0,1].

The values for y, xmin, and xmax are broadcast.

A current plot must be made to add to, as the extent of the lines is taken from the current plot.

BinderPlots.hspan! — Method

hspan!([p::Plot], ys; kwargs...)
hspan!([p::Plot], ys, YS; xmin=0.0, ymin=1.0, kwargs...)

Draw horizontal rectanglular rectangle from ys to YS. By default extends over x range of plot p, though using xmin or xmax can adjust that. These are values in [0,1] and are interpreted relative to the range returned by extrema(p).x.

If just ys is specified, it is taken as zipped form of (ys, YS). This form is from Plots.jl so use this for compatibility.

BinderPlots.image! — Method

image!([p::Plot], img_url; [x],[y],[sizex],[sizey], kwargs...)

Plot image, by url, onto background of plot.

x,y,sizex, sizey specify extent via [x, x+sizex] × [y-sizey, y].

pass sizing="stretch" to fill space.

other arguments cf. plotly examples.

Example

img = "https://upload.wikimedia.org/wikipedia/commons/thumb/1/1f/Julia_Programming_Language_Logo.svg/320px-Julia_Programming_Language_Logo.svg.png"
plot(;xlims=(0,1), ylims=(0,1), legend=false);
image!(img; sizing="stretch")
plot!(x -> x^2; linewidth=10, linecolor=:black)

BinderPlots.label! — Method

label!(str, x, y, args...)

A convenience when adding a single annotation:

x,y are coordinates

label is formed by text(str, args...)

Example

plot(sin, 0, pi)
BinderPlots.label!("hi there", 2, 0.9, :bottom, :right, 20)

BinderPlots.legend! — Method

legend!([p::Plot], legend::Bool) hide/show legend

BinderPlots.parallelogram — Method

parallelogram(q, v̄, w̄; kwargs...)
parallelogram!([p::Plot], q, v̄, w̄; kwargs...)

Plot parallelogram formed by two vectors, v̄ and w̄, both anchored at point q.

BinderPlots.plot! — Method

plot!([p::Plot], x, y; kwargs...)
plot!([p::Plot], f, a, [b]; kwargs...)
plot!([p::Plot], f; kwargs...)

Used to add a new trace to an existing plot. Like Plots.plot!. See plot for argument details.

BinderPlots.plot — Method

plot(x, y, [z]; kwargs...)
plot(f::Function, a, [b]; kwargs...)
plot(pts; kwargs...)
plot(x,[y],[z]; seriestype::Symbol=:lines, kwargs...)

Create a plot, defaulting to a line plot.

Returns a Plot instance from PlotlyLight

For lines:

x,y,[z] points to plot. NaN values break the line. Can be specified through a container, pts of $(x,y)$ or $(x,y,z)$ values.

a, b: the interval to plot a function over can be given by two numbers or if just a then by extrema(a).

label in legend

linecolor: color of line

linewidth: width of line

The line argument allows for magic arguments.

Other keyword arguments include size=(width=X, height=Y), xlims and ylims, legend, aspect_ratio.

Provides an interface like Plots.plot for plotting a function f using PlotlyLight. This just scratches the surface, but PlotlyLight allows full access to the underlying JavaScript library library.

The provided "Plots-like" functions are plot, plot!, scatter!, scatter, annotate!, title!, xlims! and ylims!.

Example

p = plot(sin, 0, 2pi; legend=false)
plot!(cos)
# add points
x0 = [pi/4, 5pi/4]
scatter!(x0, sin.(x0), markersize=10)
# add text
annotate!(tuple(zip(x0, sin.(x0), (text("A",:top), text("B",:top)))...), halign="left", pointsize=12)
title!("Sine and cosine and where they intersect in [0,2π]")
# adjust limits
ylims!((-3/2, 3/2))
# add shape
y0, y1 = extrema(p).y
[rect!(xᵢ-0.1, xᵢ+0.1, y0, y1, fillcolor="gray", opacity=0.2) for xᵢ ∈ x0]
# display plot
p

Warning

You may need to run the first plot cell twice to see an image.

BinderPlots.plot_implicit — Method

plot_implicit(f; xlims=(-5,5), ylims=(-5,5), legend=false, linewidth=2, kwargs...)
plot_implicit!([p::Plot], f; kwargs...)

For f(x,y) = ... plot implicitly defined y(x) from f(x,y(x)) = 0 over range specified by xlims and ylims.

Example

f(x,y) = x * y - (x^3 + x^2 + x + 1)
plot_implicit(f)

(Basically just contour plot with levels=0 and points determined by extrema of xlims and ylims.)

BinderPlots.poly — Method

poly(points; kwargs...)
poly!([p::Plot], points; kwargs...)

Plot polygon described by points, a container of x-y or x-y-z values. Alternative to creating Shape instance.

Example

f(r,θ) = (r*cos(θ), r*sin(θ))
poly(f.(repeat([1,2],5), range(0, 2pi-pi/5, length=10)))

BinderPlots.quiver! — Function

quiver!([p::Plot], x, y, txt=nothing; quiver=(dx, dy), kwargs...)
quiver(x, y, txt=nothing; quiver=(dx, dy), kwargs...)

Draw 2d arrows. See BinderPlots.arrow! for a single arrow.

(x,y) are tail positions, optionally labeled by txt

quiver specifies vector part of arrow

kwargs processarrowhead::Int?,arrowwidth::Int?,arrowcolor`

Example

ts = range(0, 2pi, length=100)
p = plot(sin.(ts), cos.(ts), linecolor="red")
ts = range(0, 2pi, length=10)
quiver!(p, cos.(ts), sin.(ts), quiver=(-sin.(ts), cos.(ts)), arrowcolor="red")
p

This example shows how text can be rotated with angles in degrees and positive angles measured in a clockwise direction.

ts = range(0, 2pi, 100)
p = plot(cos.(ts), sin.(ts), linecolor="red", aspect_ratio=:equal,
    linewidth=20, opacity=0.2)

txt = split("The quick brown fox jumped over the lazy dog")
ts = range(0, 360, length(txt)+1)[2:end]
for (i,t) ∈ enumerate(reverse(ts))
    quiver!(p, [cosd(t)],[sind(t)],txt[i],
            quiver=([0],[0]),
            textangle=90-t,
            font=(size=20,))
end
xaxis!(zeroline=false); yaxis!(zeroline=false) # remove zerolines
p

3d arrows

3d arrows are possible using arrows!.

BinderPlots.rect! — Method

rect!([p::Plot], x0, x1, y0, y1; kwargs...)

Draw rectangle shape on graphic.

Example [2, 3] × [-1, 1]

rect!(p, 2, 3, -1, 1; linecolor=:gray, fillcolor=:red, opacity=0.2)

BinderPlots.rgb — Function

rgb(r,g,b,α=1.0)
rgb(c::Union{RGB, RGBA}) # RGB[A] from Colors.jl
rgb(::Symbol, α)
colormap(cname, N; kwargs...)

Specify red, green, blue values between 0 and 255 (as integers). The transparency is specified by the 4th argument, a value in [0.0,1.0].

The range operator can be used with color to produce a sequence, following Colors.range for RGB[A] values. (It is not lazy, so don't take length to be too large.)

The colormap function returns a colormap of length 10 using Colors.colormap.

BinderPlots.rotate — Function

rotate(s::Shape, θ, c=center(s))
rotate!(s::Shape, θ, c=center(s))

Rotate shape about its center ccw by angle θ

BinderPlots.scale — Function

scale(s::Shape, x, y=x)
scale!(s::Shape, x, y=x)

Scale in x and y direction

BinderPlots.scatter! — Method

scatter!([p::Plot], x, y, [z]; kwargs...)
scatter!([p::Plot], pts; kwargs...)

Add points to a plot.

markershape: shape, e.g. "diamond" or "diamond-open"

markercolor: color e.g. "red"

markersize: size, as an integer

BinderPlots.scatter — Method

scatter(x, y, [z]; [markershape], [markercolor], [markersize], kwargs...)
scatter(pts; kwargs...)
scatter!([p::Plot], x, y, [z]; kwargs...)
scatter!([p::Plot], pts; kwargs...)

Place points on a plot.

markershape: shape, e.g. "diamond" or "diamond-open"

markercolor: color e.g. "red", or :blue, or "rgba(255, 0, 0, 0.5)" (a string,a sit passes JavaScript command to plotly.

markersize: size, as an integer

The marker keyword has some magic arguments.

BinderPlots.scroll_zoom! — Method

scrollzoom!([p], x::Bool) turn on/off scrolling to zoom

BinderPlots.size! — Method

size!([p::Plot]; [width], [height]) specify size of plot figure

BinderPlots.skirt — Method

skirt(q, v, f::Function; kwargs...)
skirt!([p::Plot], q, v, f::Function; kwargs...)
skirt!([p::Plot], xs, ys, zs, f::Function; kwargs...)

Along a path (xs, ys, zs) plot a skirt between the path and (xs, ys, f(xs, ys)). The case of a path described by a vector, v, anchored at a point q has a special method.

Example

x(r, θ) = r*cos(θ)
y(r, θ) = r*sin(θ)
f(x,y) = 4 - x^2 - y^2
rs, θs = range(0,2,length=10), range(0, 2pi, length=20)
xs, ys = x.(rs', θs), y.(rs', θs)
zs = f.(xs, ys)
surface(xs, ys, zs, opacity=.2, showscale=false)

q = [0,0,0]
v = [2,0,0]
d = skirt!(q, v, f; opacity=0.6)

t = range(0, 1, length=100)
xs = 2 * t .* sin.(t*pi/2)
ys = 2 * t .* cos.(t*pi/2)
zs = zero.(xs)
skirt!(xs, ys, zs, f; color="blue", opacity=0.6)
skirt!(xs, -ys, zs, f; color="blue", opacity=0.6)

BinderPlots.stroke — Method

stroke(args...; alpha=nothing)

Specify line attributes. Returns a Stroke instance.

This function can be used to specify the attributes of the polygonal line in a shape or other uses of lines.

BinderPlots.surface — Method

surface(x, y, z; kwargs...)
surface!(x, y, z; kwargs...)
surface(x, y, f::Function; kwargs...)
surface!(x, y, f::Function; kwargs...)

Create surface plot. Pass zcontour=true to add contour plot projected onto the z axis.

Example

f(x,y) = 4 - x^2 - y^2
xs = ys = range(-2, 2, length=50)
surface(xs, ys, f; palette=cgrad(:acton))  # zs = f.(xs', ys); can use seriescolor for palette

Extended help

From https://discourse.julialang.org/t/3d-surfaces-time-slider/109673

z1 = Vector[[8.83, 8.89, 8.81, 8.87, 8.9, 8.87],
                       [8.89, 8.94, 8.85, 8.94, 8.96, 8.92],
                       [8.84, 8.9, 8.82, 8.92, 8.93, 8.91],
                       [8.79, 8.85, 8.79, 8.9, 8.94, 8.92],
                       [8.79, 8.88, 8.81, 8.9, 8.95, 8.92],
                       [8.8, 8.82, 8.78, 8.91, 8.94, 8.92],
                       [8.75, 8.78, 8.77, 8.91, 8.95, 8.92],
                       [8.8, 8.8, 8.77, 8.91, 8.95, 8.94],
                       [8.74, 8.81, 8.76, 8.93, 8.98, 8.99],
                       [8.89, 8.99, 8.92, 9.1, 9.13, 9.11],
                       [8.97, 8.97, 8.91, 9.09, 9.11, 9.11],
                       [9.04, 9.08, 9.05, 9.25, 9.28, 9.27],
                       [9, 9.01, 9, 9.2, 9.23, 9.2],
                       [8.99, 8.99, 8.98, 9.18, 9.2, 9.19],
                       [8.93, 8.97, 8.97, 9.18, 9.2, 9.18]]
xs , ys = 1:length(z1[1]), 1:length(z1) # needed here given interface chosen
surface(xs, ys, z1, colorscale="Viridis")
surface!(xs, ys, map(x -> x .+ 1, z1), colorscale="Viridis", showscale=false, opacity=0.9)
surface!(xs, ys, map(x -> x .- 1, z1), colorscale="Viridis", showscale=false, opacity=0.9)

Julia users would typically use a matrix to hold the z data, but Javascript users would expect a vector of vectors, as above. As PlotlyLight just passes on the data to Javascript, the above is perfectly acceptable. Indeed BinderPlots converts matrices to this format in such plots.

(The keyword arguments above come from Plotly, not Plots.)

A parameterized surface can be displayed. Below the unexported unzip function returns 3 matrices specifying the surface described by the vector-valued function r.

import BinderPlots: unzip
r1, r2 = 2, 1/2
r(u,v) = ((r1 + r2*cos(v))*cos(u), (r1 + r2*cos(v))*sin(u), r2*sin(v))
us = vs = range(0, 2pi, length=25)
xs, ys, zs = unzip(us, vs, r)

surface(xs, ys, zs)

Plotting a vertical plane poses a slight challenge, as we can't parameterize as z=f(x,y). Here we intersect a surface with the plane ax + by + 0z = d and add a trace for the intersection of the two surfaces.

f(x, y) = 4 - (x^2 + y^2)

# surface
xs =  ys = range(-2, 2, length=100)
zs = f.(xs', ys)

# (vertical) plane
m, M = extrema(zs)
zzs = range(m, M, length=2)
Xs = ((x,z) -> x).(xs', zzs)
Zs = ((x,z) -> z).(xs', zzs)
a, b, c, d = 1, 1, 0, 1
plane(x,z) = (d - a*x - c*z) / b
Ys = plane.(xs', zzs)

# intersection
y(t) = (d - a*t) / b
γ(t) = (t, y(t), f(t, y(t)))

surface(xs, ys, zs)
surface!(Xs, Ys, Zs, opacity=0.2)
plot!(unzip(γ.(xs))...; linewidth=3)

# single transparent color
cs =  cgrad([parse(Colorant, :seagreen)], alpha=0.25)
surface(xs, ys, zs; seriescolor=cs)

BinderPlots.text — Method

text(str, args...; kwargs...)
text(str, f::Font)

Create text with font information to be passed to labeling functions.

f::Font: object produced by font

args..., kwargs...: passed to font to create font information. The positional arguments are matched by type.

BinderPlots.title! — Method

title!([p::Plot], txt)
xlabel!([p::Plot], txt)
ylabel!([p::Plot], txt)
zlabel!([p::Plot], txt)

Set plot title.

txt: either a string or a string with font information produced by text.

Example

f = font(20, :red, :left, :bottom, 45.0)

p = plot(sin, 0, 2pi)
title!(p, text("Plot", f))
xlabel!(p, text("sine function", f))
xticks!(p, 0:pi:2pi, ticklabels = ["0", "π", "2π"], tickfont=f)
annotate!(p, [(pi, 0, text("Zero", f))])
quiver!(p, [pi],[1/2], [text("zero",f)], quiver=([0],[-1/2]))

BinderPlots.translate — Function

translate(s::Shape, x, y=x)
translate!(s::Shape, x, y=x)

Shift shape over by x, up by y

BinderPlots.unzip — Method

unzip(v, [vs...])
unzip(f::Function, a, b)
unzip(a, b, F::Function)

Reshape data to x,y,[z] mode.

In its basic use, zip takes two vectors, pairs them off, and returns an iterator of tuples for each pair. For unzip a vector of same-length vectors is "unzipped" to return two (or more) vectors.

The function version applies f to a range of points over (a,b) and then calls unzip. This uses the adapted_grid function from PlotUtils.

The function version with F computes F(a', b) and then unzips. This is used with parameterized surface plots

This uses the invert function of SplitApplyCombine.

BinderPlots.vline! — Method

vline!(x; ymin=0, ymax=1.0; kwargs...)

Draw vertical line at x. By default extends over the current plot range, this can be adjusted by ymin and ymax, values in [0,1].

The values for x, ymin, and ymax are broadcast.

A current plot must be made to add to, as the extent of the lines is taken from that.

Example

Add a grid to a plot:

p = plot(x -> x^2, 0, 1; aspect_ratio=:equal)
vline!(0:.1:1, linecolor=:red,  opacity=0.25, linewidth=5)
hline!(0:.1:1, line=(:blue, 0.75))

BinderPlots.vspan! — Method

vspan!([p::Plot], xs; kwargs...)
vspan!([p::Plot], xs, XS; ymin=0.0, ymin=1.0, kwargs...)

Draw vertical rectanglular rectangle from xs to XS. By default extends over y range of plot p, though using ymin or ymax can adjust that. These are values in [0,1] and are interpreted relative to the range returned by extrema(p).y.

If just xs is specified, it is taken as zipped form of (xs, XS). This form is from Plots.jl so use this for compatibility.

Example

p = plot(x -> x^2, 0, 1; legend=false)
M = 1 # max of function on `[a,b]`
vspan!(0:.1:0.9, 0.1:0.1:1.0; ymax=[x^2 for x in 0:.1:0.9]/M,
    fillcolor=:red, opacity=.25)

BinderPlots.wireframe — Method

wireframe(x, y, z; kwargs...)
wireframe(x, y, f::Function; kwargs...)
wireframe!([p::Plot], x, y, z; kwargs...)
wireframe!([p::Plot], x, y, f::Function; kwargs...)

Create wireframe.

Example

f(x, y) = 4 - x^2 - y^2
xs = ys = range(-2, 2, length=100)
surface(xs, ys, f)
wireframe!(xs, ys, f)

BinderPlots.xaxis! — Method

xaxis!([p::Plot]; kwargs...)
yaxis!([p::Plot]; kwargs...)
zaxis!([p::Plot]; kwargs...)

Adjust properties of an axis on a chart using Plotly keywords.

ticks, ticktext, ticklen, tickwidth, tickcolor, tickfont, showticklabels

showgrid, gridcolor, gridwidth

zeroline, zerolinecolor, zerolinewidth

BinderPlots.xlabel! — Method

xlabel!([p::Plot], txt::Union{String, TextFont})
ylabel!([p::Plot], txt::Union{String, TextFont})
zlabel!([p::Plot], txt::Union{String, TextFont})

Set axis label. Use a text object to specify font information.

BinderPlots.xlims! — Method

xlims!(p, lims) set x limits of plot

BinderPlots.xticks! — Method

xticks!([p::Plot], ticks; [ticklabels], [tickfont], kwargs...)
yticks!([p::Plot], ticks; [ticklabels], [tickfont], kwargs...)
zticks!([p::Plot], ticks; [ticklabels], [tickfont], kwargs...)

Set ticks. Optionally add labels using a matching length container. Passing a Font object to tickfont will set the font.,

ticks: a range of collection of tick positions

ticklabels: if given, a matching length collection of strings

tickfont: a Font instance to adjust font of all specified ticks.

kwargs...: passed to [xyz]axis! method.

BinderPlots.ylims! — Method

ylims!(p, lims) set y limits of plot

BinderPlots.ziptie — Method

ziptie(xs, ys, zs, xs′, ys′, zs′; kwargs...)
ziptie!([p::Plot], xs, ys, zs, xs′, ys′, zs′; kwargs...)

Surface created by connecting points along two paths given by (xs, ys, zs) and (xs′, ys′, zs′). All vectors must be same length. Mesh is created by zipping together points on the two curves. Makie docs refer to this as a ruled surface.

Example

r(t) = (sin(t), cos(t), t)
s(t) = (sin(t+pi), cos(t+pi), t)
ts = range(0, 4pi, length=100)
ziptie(unzip(r.(ts))..., unzip(s.(ts))...;
       color="green", opacity=.25, showscale=false)

BinderPlots.zlims! — Method

zlims!(p, lims) set z limits of plot

BinderPlots.★ — Method

★(q, xs, ys, zs; kwargs...)
★!([p::Plot], q, xs, ys, zs; kwargs...)

A star connected region has an origin, q, for which each boundary point (described by (xs, ys, zs) is accessible by a ray for q which does not cross the boundary.

Example

pts = 5
Δ = 2pi/pts/2
a, A = 1, 3
q = [0,0,0]
ts = range(0, 2pi, length=pts+1)
ps = [(A*[cos(t),sin(t),0], a*[cos(t+Δ), sin(t+Δ), 0]) for t in ts]
xs, ys, zs = unzip(collect(Base.Iterators.flatten(ps)))
★(q, xs, ys, zs)